functions/storage/requirements.txt
firebase-admin==4.0.1
python-dateutil==2.8.1
google-cloud-pubsub==1.4.2
google-cloud-storage==1.27.0
